+++ LogicBuilder.Expressions.Utils.Tests/Expansions/ExpansionsHelperTest.cs
@@ -35,13 +35,16 @@ namespace LogicBuilder.Expressions.Utils.Tests.Expansions
         }
 
         [Fact]
-        public void GetExpansionSelectors_ReturnsEmptyListIfSelectExpandDefinitionIsNull()
+        public void GetExpansionSelectors_ReturnsListOfValueTypeSelectorsIfSelectExpandDefinitionIsNull()
         {
             //Act
             var expressions = ExpansionsHelper.GetExpansionSelectors<CategoryModel>(null);
+            var expressionsToString = expressions.Select(ExpressionStringBuilder.ToString).ToHashSet();
 
             //Assert
-            Assert.Empty(expressions);
+            Assert.NotEmpty(expressions);
+            Assert.Contains("i => Convert(i.CategoryID)", expressionsToString);
+            Assert.Contains("i => i.CategoryName", expressionsToString);
         }
 
         [Fact]